Move resource to api v1
